+++ lib/src/base/SubTypes/subType.vue
@@ -3,7 +3,8 @@
     <div>
       <div class="text-video content-wrapper d-flex" v-show="name">
         <TranslationComponent v-if="translationComponentSupport" :locales="locales"  @setFormLang="(locale) => formLang = locale"/>
-        <component :is="getComponentForm" :ref="name" :locales="locales" :selectedLang="formLang" />
+        <component :is="getComponentForm" :ref="name" :locales="locales" :selectedLang="formLang" :selected-media="selectedMedia" @openMediaModal="$refs.sectionsMediaComponent.openModal()" @closeMediaModal="$refs.sectionsMediaComponent.closeModal()" />
+        <MediaComponent ref="sectionsMediaComponent" :sections-user-id="sectionsUserId" @emittedMedia="(media) => selectedMedia = media"></MediaComponent>
       </div>
     </div>
     <button
@@ -20,9 +21,11 @@
 <script>
 import { importComp } from "../../utils";
 import TranslationComponent from "../../components/Translations/TranslationComponent";
+import MediaComponent from "../../components/Medias/MediaComponent.vue";
 
 export default {
   components: {
+    MediaComponent,
     TranslationComponent
   },
   props: {
@@ -39,12 +42,17 @@ export default {
     translationComponentSupport: {
       type: Boolean,
       default: false
+    },
+    sectionsUserId: {
+      type: String,
+      default: ''
     }
   },
   data() {
     return {
       withTabs: false,
-      formLang: this.locales[0]
+      formLang: this.locales[0],
+      selectedMedia: null
     };
   },
   computed: {